Add ordering for single client
When an item is checked, it will move to the bottom of the live list, and back to the top when it is unchecked. So all the completed items should always live at the bottom of the live list. Now need to add sending the update to the other clients